j'ai corriger l'arabe

+++ backend/src/main/java/fr/projet/kitcinq/studentcall/StudentCallService.java
@@ -1,7 +1,7 @@
 package fr.projet.kitcinq.studentcall;
 
 public interface StudentCallService {
-    void launch(int courseId);  // Lancer l'appel pour un cours
-    void setStudentPresent(int courseId, int studentId);  // Marquer un étudiant présent
-    void setStudentAbsent(int courseId, int studentId);  // Marquer un étudiant absent
+    void launch(long courseId);
+    void setStudentPresent(long courseId, long studentId);
+    void setStudentAbsent(long courseId, long studentId);
 }
